Tests/LibWeb/Text/expected/css-accent-color.txt

Before bgColor=rgb(105, 105, 105)
After bgColor=rgb(0, 0, 255)
